+++ packages/zowe-explorer-api/src/profiles/ZoweExplorerZosmfApi.ts
@@ -60,7 +60,7 @@ export namespace ZoweExplorerZosmf {
         }
 
         private _getSession(serviceProfile: imperative.IProfileLoaded): imperative.Session {
-            let cmdArgs: imperative.ICommandArguments = {
+            const cmdArgs: imperative.ICommandArguments = {
                 $0: "zowe",
                 _: [""],
                 host: serviceProfile.profile.host as string,
@@ -68,20 +68,13 @@ export namespace ZoweExplorerZosmf {
                 protocol: serviceProfile.profile.protocol as string,
                 basePath: serviceProfile.profile.basePath as string,
                 rejectUnauthorized: serviceProfile.profile.rejectUnauthorized as boolean,
+                tokenType: serviceProfile.profile.tokenType as string,
+                tokenValue: serviceProfile.profile.tokenValue as string,
+                user: serviceProfile.profile.user as string,
+                password: serviceProfile.profile.password as string,
+                certFile: serviceProfile.profile.certFile as string,
+                certKeyFile: serviceProfile.profile.certKeyFile as string,
             };
-            if (!serviceProfile.profile.tokenValue) {
-                cmdArgs = {
-                    ...cmdArgs,
-                    user: serviceProfile.profile.user as string,
-                    password: serviceProfile.profile.password as string,
-                };
-            } else {
-                cmdArgs = {
-                    ...cmdArgs,
-                    tokenType: serviceProfile.profile.tokenType as string,
-                    tokenValue: serviceProfile.profile.tokenValue as string,
-                };
-            }
             return this.getSessionFromCommandArgument(cmdArgs);
         }
 
@@ -431,8 +424,8 @@ export namespace ZoweExplorerZosmf {
             return zosjobs.DownloadJobs.downloadSpoolContentCommon(this.getSession(), parms);
         }
 
-        public getSpoolContentById(jobname: string, jobid: string, spoolId: number): Promise<string> {
-            return zosjobs.GetJobs.getSpoolContentById(this.getSession(), jobname, jobid, spoolId);
+        public getSpoolContentById(jobname: string, jobid: string, spoolId: number, encoding?: string): Promise<string> {
+            return zosjobs.GetJobs.getSpoolContentById(this.getSession(), jobname, jobid, spoolId, encoding);
         }
 
         public getJclForJob(job: zosjobs.IJob): Promise<string> {